chore: move fileparser result into separate file
